+++ wormhole-features/pom.xml
@@ -31,5 +31,6 @@
 
     <modules>
         <module>wormhole-pipeline</module>
+        <module>wormhole-jdbc</module>
     </modules>
 </project>
