+++ wormhole-features/wormhole-jdbc/pom.xml
@@ -21,24 +21,17 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <parent>
-        <artifactId>wormhole-kernel</artifactId>
+        <artifactId>wormhole-features</artifactId>
         <groupId>com.zergclan</groupId>
         <version>0.1.0-SNAPSHOT</version>
     </parent>
-
-    <artifactId>wormhole-metadata</artifactId>
+    <artifactId>wormhole-jdbc</artifactId>
     <name>${project.artifactId}</name>
 
     <dependencies>
         <dependency>
-            <groupId>com.zergclan</groupId>
-            <artifactId>wormhole-common</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>com.zergclan</groupId>
-            <artifactId>wormhole-jdbc</artifactId>
-            <version>${project.version}</version>
+            <groupId>com.zaxxer</groupId>
+            <artifactId>HikariCP</artifactId>
         </dependency>
     </dependencies>
 </project>
