+++ wormhole-infra/wormhole-common/src/main/java/com/zergclan/wormhole/common/configuration/initializer/TargetConfigurationInitializer.java
@@ -15,34 +15,34 @@
  * limitations under the License.
  */
 
-package com.zergclan.wormhole.common.swapper;
+package com.zergclan.wormhole.common.configuration.initializer;
 
+import com.zergclan.wormhole.common.WormholeInitializer;
 import com.zergclan.wormhole.common.configuration.DataNodeConfiguration;
-import com.zergclan.wormhole.common.configuration.SourceConfiguration;
-import com.zergclan.wormhole.common.yaml.YamlSourceConfiguration;
+import com.zergclan.wormhole.common.configuration.TargetConfiguration;
+import com.zergclan.wormhole.common.configuration.yaml.YamlTargetConfiguration;
+import com.zergclan.wormhole.common.constant.MarkConstant;
+import com.zergclan.wormhole.common.util.StringUtil;
 
+import java.util.Collection;
 import java.util.LinkedHashMap;
 import java.util.Map;
 
 /**
- * YAML source configuration swapper.
+ * Initializer of {@link TargetConfiguration}.
  */
-public final class YamlSourceConfigurationSwapper implements Swapper<YamlSourceConfiguration, SourceConfiguration> {
+public final class TargetConfigurationInitializer implements WormholeInitializer<YamlTargetConfiguration, TargetConfiguration> {
     
     @Override
-    public SourceConfiguration swapToTarget(final YamlSourceConfiguration yamlConfiguration) {
+    public TargetConfiguration init(YamlTargetConfiguration yamlConfiguration) {
         String dataSource = yamlConfiguration.getDataSource();
-        String actualSql = yamlConfiguration.getActualSql();
         String table = yamlConfiguration.getTable();
-        String conditionSql = yamlConfiguration.getConditionSql();
+        Collection<String> uniqueNodes = StringUtil.deduplicateSplit(yamlConfiguration.getUniqueNodes(), MarkConstant.COMMA);
+        Collection<String> compareNodes = StringUtil.deduplicateSplit(yamlConfiguration.getCompareNodes(), MarkConstant.COMMA);
+        Collection<String> ignoreNodes = StringUtil.deduplicateSplit(yamlConfiguration.getIgnoreNodes(), MarkConstant.COMMA);
+        String versionNode = yamlConfiguration.getVersionNode();
         Map<String, DataNodeConfiguration> dataNodes = new LinkedHashMap<>();
         yamlConfiguration.getDataNodes().forEach((key, value) -> dataNodes.put(key, new DataNodeConfiguration(key, value.getNodeType(), value.getDataType(), value.getDefaultValue())));
-        return new SourceConfiguration(dataSource, actualSql, table, conditionSql, dataNodes);
-    }
-    
-    @Override
-    public YamlSourceConfiguration swapToSource(final SourceConfiguration configuration) {
-        // TODO init yamlSourceConfiguration
-        return null;
+        return new TargetConfiguration(dataSource, table, uniqueNodes, compareNodes, ignoreNodes, versionNode, dataNodes);
     }
 }
